graphics: Use float values for vertices

+++ internal/graphics/image.go
@@ -95,7 +95,7 @@ func (i *Image) Fill(clr color.RGBA) error {
 	return nil
 }
 
-func (i *Image) DrawImage(src *Image, vertices []int16, clr affine.ColorM, mode opengl.CompositeMode) error {
+func (i *Image) DrawImage(src *Image, vertices []float32, clr affine.ColorM, mode opengl.CompositeMode) error {
 	c := &drawImageCommand{
 		dst:      i,
 		src:      src,
